Projects: linux_bd: fixed board repo paths
Since we distribute the project file (.xpr) with this project, it will
not have a valid BoardPartRepoPaths parameter set that links to an
external board repository because that option is set at the time the
project is created. This means that on systems that use set_param
board.repoPaths in init.tcl to link to board files outside of the Vivado
installation directory the tools will not be able to see the board files
kept there. To resolve this I added a copy of the needed board files in
the repo folder, and set BoardPartRepoPaths to relatively point to it. A
better solution may be to revert to using create_project.tcl so that the
project is generated with this value correctly set.