Projects: linux_bd: migrated to 2016.1
Updated project to 2016.1. SDK projects all build, but have not been run
on hardware yet. The following was changed:

-An additional axi_gpio had to be used for the upper shield pins. This
is due to a bug in the axi_gpio core, where I/O buses defined in board
files cannot be connected to the second channel.
-The XADC interrupt ID was changed
-The SDK BSP and projects all had to be recreated to work in SDK 2016.1.
They should all be functionally equivalent.